Add support for building standalone executables (#2)
This commit adds a build.sh script that uses cx_Freeze to create a directory with an executable and the libraries it needs. We should be able to just zip this directory up and send it to the user. A script to create this package will be added in a future PR.