Deal with duplicate errors more gracefully (#9)
The error list is now an error set to prevent the user being inundated
by a ton of repeated instances caused by the same error. Instead, if
an identical duplicate is detected, a single "note" error is added to
the error set to indicate there were multiple occurrences of this error.

Duplicate errors are stored in their own class, rather than being
created as another instance of the class of error being duplicated.