Add support for building with PyInstaller (#18)
Something seems to be broken in the interaction between SciPy, Cython,
and cx_Freeze. It does look like there's a potential fix in the Cython
pipeline, but who knows when that will make it into a release (and when
SciPy will have a new release compiled with a new Cython). Until then,
adding an alternative PyInstaller script so that we can still create
standalone executables.