procedures: Build procedure/suite dicts in __init__
Previously, a Procedure expected to receive steps as a dict of the form:

{'s1': s1, 's2': s2}

where s1 and s2 are ProcedureStep objects with step_id equal to 's1' and
's2', respectively.

This is unnecessarily verbose and makes it easier to accidentally
construct a procedure where the step keys do not correspond to the step
IDs. Instead, pass in a list [s1, s2], and have the Procedure construct
the dict by iterating over that list. This ensures that the keys will
always be equal to the step IDs.

ProceduresEngine did something similar with the procedure suite; this
has also been changed.