proc: Add ProcLang parser
Add the first revision of the Procedures Language grammar and a parser
for converting ProcLang documents into Topside ProcedureSuites.

This version of ProcLang is fairly rudimentary and is still missing some
features that we will eventually want: logical AND/OR conditions,
tolerance in equality comparison, in-document specification of procedure
suite metadata. However, this should still be good enough for a minimum
viable product.