proc: Fix a few bugs in procedure equality (#59)
Should have added more unit tests the first time around!

Fix the following bugs:
- Equality comparisons were not checking the type of `other`, except for
in the Comparison class. This meant that comparing a Procedure to `None`
would just error, and comparing an And to an Or with the same
subconditions would be `True`!
- Comparison base class instances with custom comparison functions would
always compare equal, regardless of what the custom function is.
Unfortunately, there isn't a trivial way to check whether two lambdas
are equal to each other. For now, I've just removed the custom comp_fn
functionality, since we don't use it anywhere anyways.