proc: Add logical expressions in comparisons
So far this supports and, or, and parentheses. And is taken as having
higher precedence, so "x and y or z" is parsed as "(x and y) or z".

There are three terminal symbols for and and or symbols, "AND, "and",
and "&&". I'm not sure if that's the desired behaviour.