pdl: Fix a bug in typedef validation
Typedef validation in `File` was comparing two values that are not, in
general, expected to be equal: the number of typedef parameters
(variable) and the number of fields in a typedef entry (at least 3).
Unit tests didn't catch this because these numbers happened to be equal
for all of the examples. Fixed the comparison and added a typedef with 2
parameters to the example PDL file, which caused the existing tests to
fail before the fix.