app: Integrate plumbing engine and PDL loading
Adds the ability to load a plumbing engine from a PDL file and have it
visualized in the plumbing pane. The plumbing engine is connected to the
procedures engine and is affected by actions in the procedure. Also
adds time controls to the UI. Currently, only "step forward" and
"advance" do anything (step forward steps time by 100ms and advance runs
the plumbing engine to steady-state with a 30s timeout).