pdl: Patch a bug causing the application to hang
In the built application, the PDL import directory does not exist. This
causes Package instantiation to throw a FileNotFoundError, which hangs
the application. For now, since none of our PDL uses the import
directory, patch this by ignoring the error and raising a warning
instead. This only suppresses the error in the event that the import
directory is unused; if we actually need to import something, we will
get an error further down when we can't find that file. This is a
temporary workaround until we get dynamic import paths implemented.