app: Rewrite base application to be Widget-centric (#103)
There are two main ways to structure a Qt application: using QML or
using Qt Widgets in pure Python/C++. QML is simpler and more intuitive
for building GUIs, but it's less flexible and there isn't any way to
include Qt Widgets in a QML scene.

This commit removes the main application layout previously defined in
application.qml and instead defines an equivalent application using
pure Python. The QML used for the Plumbing and Procedures panes is used
unchanged in this new application; it's just loaded into separate
widgets by Python instead of being loaded by parent QML.

The main benefit of this change is flexibility; we can now use Qt Widget
functionality that doesn't exist in QML. In particular, this opens the
door to using widgets from the pyqtgraph package for DAQ plotting.