proc: Introduce pre- and post-nodes (#112)
Distinguish between being "after" a step from being "before" the
subsequent procedure step. This allows the procedures engine to
correctly execute the first step in a procedure.

Right now, the "procedure step forward" button in the application
executes next_step, which advances the procedures engine to the
next post-node. There might be a more intuitive behaviour, but we
can do testing on that later.